fix: reset sign-out dialog busy state
