+++ frontend/src/routes/chat/ModalContainer.svelte
@@ -6,14 +6,10 @@
   import { serverRegistry } from '$lib/state/server/registry.svelte';
   import { graphqlClientManager } from '$lib/state/server/graphqlClient.svelte';
   import { getActiveServer } from '$lib/state/activeServer.svelte';
+  import SignOutDialog from './SignOutDialog.svelte';
 
   const activeInstanceId = $derived(getActiveServer());
   const serverSegment = $derived(serverIdToSegment(activeInstanceId));
-  const currentViewedServerId = $derived(page.params.serverId ? activeInstanceId : '');
-  const activeSignOutServer = $derived(
-    currentViewedServerId ? serverRegistry.getServer(currentViewedServerId) : undefined
-  );
-  const canSignOutCurrentServer = $derived(Boolean(activeSignOutServer));
   import Dialog from '$lib/ui/Dialog.svelte';
   import ConfirmDialog from '$lib/ui/ConfirmDialog.svelte';
   import { Button } from '$lib/ui/form';
@@ -25,12 +21,6 @@
   import { refreshAttachmentUrlsForMessage } from '$lib/attachments/attachmentUrls';
   import { toast } from '$lib/ui/toast';
   import { clearLastRoom } from '$lib/storage/lastRoom';
-  import { notifyLogout } from '$lib/auth/sessionChannel';
-  import {
-    hardRedirectAfterSignOut,
-    signOutServer,
-    signOutServers
-  } from '$lib/auth/signOut';
 
   /** Get the GraphQL client for the currently active instance (derived from URL). */
   function getActiveClient() {
@@ -51,8 +41,6 @@
   let deletingMessage = $state(false);
   let deletingLinkPreview = $state(false);
   let deletingAttachment = $state(false);
-  let signingOutCurrent = $state(false);
-  let signingOutAll = $state(false);
 
   // Keep the lightbox ahead of the one-hour access ticket expiry.
   const IMAGE_MODAL_URL_REFRESH_MS = 50 * 60 * 1000;
@@ -118,68 +106,6 @@
       goto(resolve('/'));
     }
     leavingServer = false;
-  }
-
-  function firstRemainingAuthenticatedServerId(excludedId: string): string | undefined {
-    const originId = serverRegistry.originServer?.id;
-    if (originId && originId !== excludedId && serverRegistry.isAuthenticated(originId)) {
-      return originId;
-    }
-
-    return serverRegistry.servers.find(
-      (server) => server.id !== excludedId && serverRegistry.isAuthenticated(server.id)
-    )?.id;
-  }
-
-  function routeToServerOrRoot(serverId: string | undefined) {
-    if (serverId) {
-      goto(resolve('/chat/[serverId]', {
-        serverId: serverIdToSegment(serverId)
-      }));
-      return;
-    }
-
-    goto(resolve('/'));
-  }
-
-  function hardNavigateToServerOrRoot(serverId: string | undefined) {
-    hardRedirectAfterSignOut(
-      serverId ? resolve('/chat/[serverId]', { serverId: serverIdToSegment(serverId) }) : '/'
-    );
-  }
-
-  async function handleSignOutCurrentServer() {
-    const signedOutServerId = currentViewedServerId;
-    const server = activeSignOutServer;
-
-    if (!server || !signedOutServerId) {
-      return;
-    }
-
-    signingOutCurrent = true;
-
-    await signOutServer(server, serverRegistry.isOriginServer(signedOutServerId)).catch(() => {});
-
-    clearLastRoom(signedOutServerId);
-
-    if (serverRegistry.isOriginServer(signedOutServerId)) {
-      serverRegistry.clearServerAuthentication(signedOutServerId);
-      notifyLogout();
-      hardNavigateToServerOrRoot(firstRemainingAuthenticatedServerId(signedOutServerId));
-    } else {
-      serverRegistry.removeServer(signedOutServerId);
-      routeToServerOrRoot(firstRemainingAuthenticatedServerId(signedOutServerId));
-    }
-  }
-
-  async function handleSignOutAllServers() {
-    signingOutAll = true;
-    await signOutServers([...serverRegistry.servers], (serverId) =>
-      serverRegistry.isOriginServer(serverId)
-    );
-    serverRegistry.removeAll();
-    notifyLogout();
-    hardRedirectAfterSignOut('/');
   }
 
   async function handleDeleteMessage(roomId: string, eventId: string) {
@@ -316,35 +242,7 @@
     <CreateRoom onroomcreated={(roomId) => handleRoomCreated(roomId)} />
   </Dialog>
 {:else if modalType === 'logout'}
-  <Dialog visible title="Sign Out" size="md" onclose={closeModal}>
-    {#snippet footer()}
-      <div class="flex flex-wrap justify-end gap-2">
-        <Button variant="secondary" onclick={closeModal}>Cancel</Button>
-        <Button
-          variant="accent"
-          loading={signingOutCurrent}
-          disabled={signingOutAll || !canSignOutCurrentServer}
-          onclick={handleSignOutCurrentServer}
-        >
-          <span class="iconify uil--sign-out-alt"></span>
-          Current Server
-        </Button>
-        <Button
-          variant="danger"
-          loading={signingOutAll}
-          disabled={signingOutCurrent && canSignOutCurrentServer}
-          onclick={handleSignOutAllServers}
-        >
-          <span class="iconify uil--signout"></span>
-          All Servers
-        </Button>
-      </div>
-    {/snippet}
-
-    <p class="text-muted">
-      Sign out of only the selected server, or disconnect every server from this client.
-    </p>
-  </Dialog>
+  <SignOutDialog onclose={closeModal} />
 {:else if modalType === 'joinRoom' && roomId}
   {#if page.state.modal?.viewerCanJoinRoom}
     <ConfirmDialog
