+++ frontend/src/routes/chat/ModalContainerButtonMock.svelte
@@ -9,12 +9,16 @@ Tiny test double for form Button used by ModalContainer specs.
   let {
     children,
     disabled = false,
+    loading = false,
     onclick
   }: {
     children: Snippet;
     disabled?: boolean;
+    loading?: boolean;
     onclick?: (e: MouseEvent) => void;
   } = $props();
 </script>
 
-<button type="button" {disabled} {onclick}>{@render children()}</button>
+<button type="button" disabled={disabled || loading} aria-busy={loading || undefined} {onclick}>
+  {@render children()}
+</button>
